Created CRUD Operations for the Customer Modulel

+++ src/main/java/com/example/FCAI/api/Repositories/OrderRepo.java
@@ -23,7 +23,7 @@ public class OrderRepo implements RepositoryService <Order> {
     }
 
     @Override
-    public Order findById(Long id) {
+    public Order findById(int id) {
         return null;
     }
